be/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# Chatbot (using Google Gemini - FREE)
google-genai>=1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
beautifulsoup4==4.12.3
lxml==5.1.0
tenacity==8.2.3

# Sentiment Analysis
transformers==4.36.0
torch==2.1.0
praw==7.7.1
cloudscraper==1.2.71
